draft mutation permission checks

+++ backend/event/mutations/DeleteEpisodeMutation.py
@@ -3,7 +3,7 @@ from graphene import ID, Boolean, List, Mutation, NonNull, ResolveInfo
 from event.models import Episode
 from event.types.EpisodeType import EpisodeType
 from graphql_app.types.LettercraftErrorType import LettercraftErrorType
-
+from source.permissions import can_edit_source
 
 class DeleteEpisodeMutation(Mutation):
     ok = Boolean(required=True)
@@ -19,6 +19,13 @@ class DeleteEpisodeMutation(Mutation):
         except Episode.DoesNotExist:
             error = LettercraftErrorType(field="id", messages=["Episode not found."])
             return cls(ok=False, errors=[error])  # type: ignore
+
+        if not can_edit_source(info.context.user, episode.source):
+            error = LettercraftErrorType(
+                field="id",
+                messages=["Not authorised to edit data related to this source"],
+            )
+            return cls(errors=[error])
 
         episode.delete()
         return cls(ok=True, errors=[])  # type: ignore
